requirements.txt
accelerate==1.3.0
docx2txt==0.8
faiss-cpu==1.10.0
fastapi==0.115.8
huggingface-hub==0.28.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.8
lxml==5.3.1
mpmath==1.3.0
packaging==24.2
pdf2image==1.17.0
Pillow==9.5.0
pypdf==5.3.0
python-docx==1.1.2
python-oxmsg==0.0.2
python-pptx==1.0.2
PyYAML==6.0.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
tokenizers==0.21.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
transformers==4.48.3
unstructured==0.16.20
unstructured-client==0.30.1
uvicorn==0.34.0
webencodings==0.5.1
XlsxWriter==3.2.2

